fix(tennis): render the API key state and stop retrying rejected keys
The widget destructured only `data` from the query, so a missing or rejected
LIVE_TENNIS_API_KEY fell through to the generic empty state. The router does
raise UNAUTHORIZED and the definition does declare a dedicated error entry, but
nothing connected the two: react-query defaults `throwOnError` to false and this
repository sets no global override, so the error never reached WidgetError. The
documented "No valid Live Tennis API key configured" message was therefore
unreachable.

Escalate only that error to the boundary, so every other failure keeps the
inline empty state, and skip the query client's three default retries for it —
a rejected key is not transient, and retrying only delayed the message by three
backoff rounds while the board showed nothing.

Also addresses review feedback: `matchCount` is now constrained to an integer in
both the router schema and the slider validator, and `getMatches` declares MCP
metadata like the other widget queries.

Adds a render test over WidgetError. The existing unit tests only proved the
predicate and the definition entry; they would have passed against the original
bug. The render test fails if the error entry is removed, in which case the
widget claims a permissions problem rather than a configuration one.

+++ packages/widgets/src/tennis/index.ts
@@ -28,7 +28,7 @@ export const { definition, componentLoader } = createWidgetDefinition("tennis", 
       }),
       matchCount: factory.slider({
         defaultValue: 5,
-        validate: z.number().min(1).max(20),
+        validate: z.number().int().min(1).max(20),
         step: 1,
         withDescription: true,
       }),
